Move even more containers, remove imports in io
